Refactor report structure
